Added input error checking

+++ app/src/main/java/com/example/passwordapp/AddPasswordFragment.kt
@@ -1,12 +1,13 @@
 package com.example.passwordapp
 
-import android.content.ClipData
 import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
+import androidx.core.view.isEmpty
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
@@ -34,18 +35,37 @@ class AddPasswordFragment : Fragment() {
     }
     private val navigationArgs: AddPasswordFragmentArgs by navArgs()
     lateinit var password: Password
-    private var passwordId = -1
+    var passwordId = -1 // // default value, didn't receive password
 
-    private var _binding: FragmentAddPasswordBinding? = null
-    private val binding get() = _binding!!
+    private lateinit var binding: FragmentAddPasswordBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        _binding = FragmentAddPasswordBinding.inflate(inflater, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_add_password, container,false)
         return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        passwordId = navigationArgs.passwordId
+        if (passwordId > 0) { // we received a password
+            viewModel.retrievePassword(passwordId).observe(viewLifecycleOwner) { selectedPassword ->
+                password = selectedPassword
+                binding.currentPassword = selectedPassword // set data binding currentPassword
+//                binding.btnDelete.isEnabled = true
+            }
+        }
+
+        binding.apply {
+            passwordViewModel = viewModel
+            lifecycleOwner = viewLifecycleOwner
+            addPasswordFragment = this@AddPasswordFragment
+            // onclick listener in xml and data binding
+        }
     }
 
     override fun onDestroyView() {
@@ -54,40 +74,13 @@ class AddPasswordFragment : Fragment() {
         val inputMethodManager = requireActivity().getSystemService(Context.INPUT_METHOD_SERVICE) as
                 InputMethodManager
         inputMethodManager.hideSoftInputFromWindow(requireActivity().currentFocus?.windowToken, 0)
-        _binding = null
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        // returns -1 (default value), or Password id
-        passwordId = navigationArgs.passwordId
-        viewModel.retrievePassword(passwordId).observe(viewLifecycleOwner) { selectedPassword ->
-            if (passwordId != -1) {
-                password = selectedPassword
-                bind(selectedPassword)
-            }
-        }
-        binding.apply {
-            // Save button onClick to save password
-            btnSave.setOnClickListener {
-                updatePassword()
-            }
-            // Cancel button to navigate back to PasswordFragment
-            btnCancel.setOnClickListener {
-                cancelPasswordCreation()
-            }
-            // Delete button onClick
-            btnDelete.setOnClickListener {
-                deletePassword()
-            }
-        }
     }
 
     /**
      * Add new password or Updates existing password
      */
-    private fun updatePassword() {
+    fun updatePassword() {
+        checkForEmptyInput()
         // Check for valid password input
         if (isEntryValid()) {
             // 1. Are we adding?
@@ -107,32 +100,49 @@ class AddPasswordFragment : Fragment() {
                     binding.tvPassword.text.toString()
                 )
             }
+            // Navigate back to password fragment
+            val action = AddPasswordFragmentDirections.actionAddPasswordFragmentToPasswordFragment()
+            findNavController().navigate(action)
         }
-        // Navigate back to password fragment
-        val action = AddPasswordFragmentDirections.actionAddPasswordFragmentToPasswordFragment()
-        findNavController().navigate(action)
+    }
+
+    /**
+     * Apply error signs if text is empty
+     */
+    private fun checkForEmptyInput() {
+        binding.apply {
+            if (tvWebsite.text.toString().isBlank()) {
+                websiteLabel.error = "*Required"
+            } else {
+                websiteLabel.error = null
+            }
+            if (tvUsername.text.toString().isBlank()) {
+                usernameLabel.error = "*Required"
+            } else {
+                usernameLabel.error = null
+            }
+            if (tvPassword.text.toString().isBlank()) {
+                passwordLabel.error = "*Required"
+            } else {
+                passwordLabel.error = null
+            }
+        }
     }
 
     /**
      * Delete password
      */
-    private fun deletePassword() {
-        viewModel.deleteItem(password)
-        findNavController().navigateUp()
-    }
-
-    /**
-     * Prepopulate text views with this password's information
-     */
-    private fun bind(password: Password) {
-        binding.apply {
-            tvWebsite.setText(password.websiteName)
-            tvUsername.setText(password.username)
-            tvPassword.setText(password.password)
+    fun deletePassword() {
+        if (isEntryValid()) {
+            viewModel.deleteItem(password)
+            findNavController().navigateUp()
         }
     }
 
-    private fun cancelPasswordCreation() {
+    /**
+     * Navigate back to home page
+     */
+    fun cancelPasswordCreation() {
         findNavController().navigate(R.id.action_addPasswordFragment_to_passwordFragment)
     }
